+++ camera/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
+## 0.3.0 (7. May, 2024)
+### Changed
+- (Breaking) changed flat control to use translation Vec3.
+
+### Added
+- Added inputs(), inputs_mut(), settings() and settings_mut() to controls.
+
 ## 0.2.0 (6. May, 2024)
 ### Changed
 - (Breaking)  removed mint types so its all done directly.